requirements.txt
tensorflow
matplotlib
streamlit
streamlit-drawable-canvas
numpy
pillow
opencv-python-headless
